bootenv: move NewPrimaryKey to keydata_test.go
