+++ bootenv/keydata_test.go
@@ -36,6 +36,7 @@ import (
 	"golang.org/x/xerrors"
 	. "gopkg.in/check.v1"
 
+	"github.com/snapcore/secboot"
 	. "github.com/snapcore/secboot"
 	. "github.com/snapcore/secboot/bootenv"
 	"github.com/snapcore/secboot/internal/testutil"
@@ -49,6 +50,16 @@ type keyDataPlatformSuite struct {
 }
 
 var _ = Suite(&keyDataPlatformSuite{})
+
+func NewPrimaryKey(sz1 int) (secboot.PrimaryKey, error) {
+	primaryKey := make(secboot.PrimaryKey, sz1)
+	_, err := rand.Read(primaryKey)
+	if err != nil {
+		return nil, err
+	}
+
+	return primaryKey, nil
+}
 
 func (s *keyDataPlatformSuite) TestNewKeyDataScopeSuccess(c *C) {
 	primaryKey, err := NewPrimaryKey(32)
